# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/nestle/overview": {
      "get": {
        "operationId": "AdminNestle_overview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Nestlé real-time overview KPIs",
        "tags": [
          "Admin Nestlé dashboard"
        ]
      }
    },
    "/admin/nestle/export.pdf": {
      "get": {
        "operationId": "AdminNestle_exportPdf",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Nestlé programme PDF report pack",
        "tags": [
          "Admin Nestlé dashboard"
        ]
      }
    },
    "/admin/nestle/visits-map": {
      "get": {
        "operationId": "AdminNestle_visitsMap",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Visit map points",
        "tags": [
          "Admin Nestlé dashboard"
        ]
      }
    },
    "/admin/nestle/export.csv": {
      "get": {
        "operationId": "AdminNestle_exportCsv",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "CSV export of visits",
        "tags": [
          "Admin Nestlé dashboard"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nestlé Ghana Vendor Onboarding API",
    "description": "API documentation for Nestlé Ghana vendor onboarding",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}